Added router & seperated email controllers
