+++ server/router.js
@@ -1,0 +1,8 @@
+const router = require('express').Router();
+const { emailNewUser } = require('./controllers/gmail.js');
+
+router
+	.route('/confirmationEmail')
+	.post(emailNewUser)
+
+module.exports = router;
